+++ tests/test_prepare.py
@@ -114,7 +114,7 @@ class TestPrepare:
 
 
 @patch('nemo_cmd.prepare.logger')
-@patch('nemo_cmd.prepare._remove_run_dir')
+@patch('nemo_cmd.prepare.remove_run_dir')
 class TestGetRunDescValue:
     """Unit tests for get_run_desc_value function.
     """
@@ -284,19 +284,19 @@ class TestRemoveRunDir:
 
     def test_remove_run_dir(self, tmpdir):
         p_run_dir = tmpdir.ensure_dir('run_dir')
-        nemo_cmd.prepare._remove_run_dir(Path(str(p_run_dir)))
+        nemo_cmd.prepare.remove_run_dir(Path(str(p_run_dir)))
         assert not p_run_dir.check()
 
     def test_remove_run_dir_file(self, tmpdir):
         p_run_dir = tmpdir.ensure_dir('run_dir')
         p_run_dir.ensure('namelist')
-        nemo_cmd.prepare._remove_run_dir(Path(str(p_run_dir)))
+        nemo_cmd.prepare.remove_run_dir(Path(str(p_run_dir)))
         assert not p_run_dir.join('namelist').check()
         assert not p_run_dir.check()
 
     @patch('nemo_cmd.prepare.os.rmdir')
     def test_remove_run_dir_no_run_dir(self, m_rmdir):
-        nemo_cmd.prepare._remove_run_dir(Path('run_dir'))
+        nemo_cmd.prepare.remove_run_dir(Path('run_dir'))
         assert not m_rmdir.called
 
 
@@ -943,29 +943,25 @@ class TestMakeExecutableLinks:
         assert p_run_dir.join('xios_server.exe').check(file=True, link=True)
 
 
+@patch('nemo_cmd.prepare.logger')
+@patch('nemo_cmd.prepare.remove_run_dir')
 class TestMakeGridLinks:
     """Unit tests for `nemo prepare` _make_grid_links() function.
     """
 
-    @patch('nemo_cmd.prepare._remove_run_dir')
-    @patch('nemo_cmd.prepare.logger')
-    def test_no_grid_coordinates_key(self, m_logger, m_rm_run_dir):
+    def test_no_grid_coordinates_key(self, m_rm_run_dir, m_logger):
         run_desc = {}
         with pytest.raises(SystemExit):
             nemo_cmd.prepare._make_grid_links(run_desc, Path('run_dir'))
         m_rm_run_dir.assert_called_once_with(Path('run_dir'))
 
-    @patch('nemo_cmd.prepare._remove_run_dir')
-    @patch('nemo_cmd.prepare.logger')
-    def test_no_grid_bathymetry_key(self, m_logger, m_rm_run_dir):
+    def test_no_grid_bathymetry_key(self, m_rm_run_dir, m_logger):
         run_desc = {'grid': {'coordinates': 'coords.nc'}}
         with pytest.raises(SystemExit):
             nemo_cmd.prepare._make_grid_links(run_desc, Path('run_dir'))
         m_rm_run_dir.assert_called_once_with(Path('run_dir'))
 
-    @patch('nemo_cmd.prepare._remove_run_dir')
-    @patch('nemo_cmd.prepare.logger')
-    def test_no_forcing_key(self, m_logger, m_rm_run_dir):
+    def test_no_forcing_key(self, m_rm_run_dir, m_logger):
         run_desc = {
             'grid': {
                 'coordinates': 'coords.nc',
@@ -976,9 +972,7 @@ class TestMakeGridLinks:
             nemo_cmd.prepare._make_grid_links(run_desc, Path('run_dir'))
         m_rm_run_dir.assert_called_once_with(Path('run_dir'))
 
-    @patch('nemo_cmd.prepare._remove_run_dir')
-    @patch('nemo_cmd.prepare.logger')
-    def test_no_link_path_absolute_coords_bathy(self, m_logger, m_rm_run_dir):
+    def test_no_link_path_absolute_coords_bathy(self, m_rm_run_dir, m_logger):
         run_desc = {
             'grid': {
                 'coordinates': '/coords.nc',
@@ -994,10 +988,8 @@ class TestMakeGridLinks:
         )
         m_rm_run_dir.assert_called_once_with(Path('run_dir'))
 
-    @patch('nemo_cmd.prepare._remove_run_dir')
-    @patch('nemo_cmd.prepare.logger')
     def test_no_link_path_relative_coords_bathy(
-        self, m_logger, m_rm_run_dir, tmpdir
+        self, m_rm_run_dir, m_logger, tmpdir
     ):
         forcing_dir = tmpdir.ensure_dir('foo')
         grid_dir = forcing_dir.ensure_dir('grid')
@@ -1020,9 +1012,7 @@ class TestMakeGridLinks:
         )
         m_rm_run_dir.assert_called_once_with(Path(str(run_dir)))
 
-    @patch('nemo_cmd.prepare._remove_run_dir')
-    @patch('nemo_cmd.prepare.logger')
-    def test_link_paths(self, m_logger, m_rm_run_dir, tmpdir):
+    def test_link_paths(self, m_rm_run_dir, m_logger, tmpdir):
         forcing_dir = tmpdir.ensure_dir('foo')
         grid_dir = forcing_dir.ensure_dir('grid')
         grid_dir.ensure('coords.nc')
@@ -1046,9 +1036,7 @@ class TestMakeGridLinks:
         assert Path(str(run_dir),
                     'bathy_meter.nc').samefile(str(grid_dir.join('bathy.nc')))
 
-    @patch('nemo_cmd.prepare._remove_run_dir')
-    @patch('nemo_cmd.prepare.logger')
-    def test_agrif_link_paths(self, m_logger, m_rm_run_dir, tmpdir):
+    def test_agrif_link_paths(self, m_rm_run_dir, m_logger, tmpdir):
         forcing_dir = tmpdir.ensure_dir('foo')
         grid_dir = forcing_dir.ensure_dir('grid')
         grid_dir.ensure('coords.nc')
@@ -1123,7 +1111,7 @@ class TestMakeForcingLinks:
         )
 
     @patch('nemo_cmd.prepare.logger')
-    @patch('nemo_cmd.prepare._remove_run_dir')
+    @patch('nemo_cmd.prepare.remove_run_dir')
     def test_no_link_path(self, m_rm_run_dir, m_log, tmpdir):
         p_nemo_forcing = tmpdir.ensure_dir('NEMO-forcing')
         run_desc = {
@@ -1173,7 +1161,8 @@ class TestMakeForcingLinks:
             Path('run_dir'), Path(p_atmos_ops), 'namelist_cfg'
         )
 
-    def test_unknown_link_checker(self, tmpdir):
+    @patch('nemo_cmd.prepare.remove_run_dir')
+    def test_unknown_link_checker(self, m_rm_run_dir, tmpdir):
         p_nemo_forcing = tmpdir.ensure_dir('NEMO-forcing')
         p_atmos_ops = tmpdir.ensure_dir(
             'results/forcing/atmospheric/GEM2.5/operational'
@@ -1193,7 +1182,6 @@ class TestMakeForcingLinks:
             }
         }
         patch_symlink_to = patch('nemo_cmd.prepare.Path.symlink_to')
-        nemo_cmd.prepare._remove_run_dir = Mock()
         with patch_symlink_to as m_symlink_to:
             with pytest.raises(SystemExit):
                 nemo_cmd.prepare._make_forcing_links(run_desc, Path('run_dir'))
@@ -1308,7 +1296,7 @@ class TestMakeRestartLinks:
         m_symlink_to.assert_called_once_with(Path(str(p_agrif_results)))
 
     @patch('nemo_cmd.prepare.logger')
-    @patch('nemo_cmd.prepare._remove_run_dir')
+    @patch('nemo_cmd.prepare.remove_run_dir')
     def test_no_link_path(self, m_rm_run_dir, m_logger, tmpdir):
         run_desc = {
             'restart': {