+++ docs/CHANGES.rst
@@ -11,6 +11,7 @@ Change Log
   * ``nemo_cmd.prepare.check_nemo_exec()``
   * ``nemo_cmd.prepare.check_xios_exec()``
   * ``nemo_cmd.prepare.make_run_dir()``
+  * ``nemo_cmd.prepare.remove_run_dir()``
 
 * Fix bugs in the ``NEMO.sh`` script generation re: extra ``{}`` in some shell
   variable expressions.